fix(core): rollback ai service (#8321)
closes #8320
